Agrega consulta de DistanciaPromedio tomando las coordenadas, en lugar del atributo Distance(mi)
Test funcionando

+++ Core/src/main/java/Repository/IAverageDistanceRepository.java
@@ -1,5 +1,5 @@
 package repository;
 
 public interface IAverageDistanceRepository {
-    Float queryAverageDistance();
+    Double queryAverageDistance();
 }
