+++ Core/src/main/java/UseCase/Input/IAverageDistanceInput.java
@@ -2,5 +2,5 @@ package usecase.input;
 
 public interface IAverageDistanceInput {
 
-    Float getAverageDistance();
+    Double getAverageDistance();
 }
